requirements.txt
requests
beautifulsoup4
Pillow
av
